gpu-server/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0

# AI (Claude Agent SDK — Max 定額プラン対応)
claude-agent-sdk>=0.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# GPU - OCR
paddlepaddle-gpu>=3.0.0
paddleocr>=2.9.0

# GPU - Vision
torch>=2.4.0
transformers>=4.45.0
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
